- rimossi fmin e fmax (sostituiti da std::min, std::max) - rimossi po' di cast (double) che servivano solo a far star zitto eclipse (che ora ovviamente si lamenta). - sostituito log2 (che non è implementato da vc++) con log(n)/log(2) dove LOG_2 è stata definita come costante

+++ source/meshing/MeshingUtils.hpp
@@ -36,8 +36,8 @@ public:
 	static bool isBorderCorner(const Eigen::Vector3d &halfExtents, const Eigen::Vector3d &corner) {
 		for (int r = 0; r < 3; ++r) {
 			if (CommonUtils::doubleEquals(
-					(double)halfExtents(r),
-					fabs((double)corner(r)))
+					halfExtents(r),
+					fabs(corner(r)))
 			) {
 				return true;
 			}
